+++ src/main/java/com/taptapsend/model/Taux.java
@@ -10,22 +10,22 @@ public class Taux {
     private String idtaux;
 
     @Column(name = "montant1")
-    private int montant1;
+    private double montant1; // CHANGÉ: int -> double
 
     @Column(name = "montant2")
-    private int montant2;
+    private double montant2; // CHANGÉ: int -> double
 
-    // Constructeurs, getters et setters
-
+    // Constructeurs
     public Taux() {
     }
 
-    public Taux(String idtaux, int montant1, int montant2) {
+    public Taux(String idtaux, double montant1, double montant2) { // CHANGÉ: int -> double
         this.idtaux = idtaux;
         this.montant1 = montant1;
         this.montant2 = montant2;
     }
 
+    // Getters et Setters
     public String getIdtaux() {
         return idtaux;
     }
@@ -34,19 +34,19 @@ public class Taux {
         this.idtaux = idtaux;
     }
 
-    public int getMontant1() {
+    public double getMontant1() { // CHANGÉ: int -> double
         return montant1;
     }
 
-    public void setMontant1(int montant1) {
+    public void setMontant1(double montant1) { // CHANGÉ: int -> double
         this.montant1 = montant1;
     }
 
-    public int getMontant2() {
+    public double getMontant2() { // CHANGÉ: int -> double
         return montant2;
     }
 
-    public void setMontant2(int montant2) {
+    public void setMontant2(double montant2) { // CHANGÉ: int -> double
         this.montant2 = montant2;
     }
 }
